Added ability to use redis' incr to increment and decrement values.
To do this, we need to keep integers untouched while pickling any other type
of value. Existing tests should prove that the native incr is working, but
I added a test to show that if an integer was pickled and native incr is used,
then the cache backend will convert the pickled integer correctly.